\name{chunk.tapply}
\alias{chunk.tapply}
\title{
chunk.tapply
}
\description{
Apply over a chunk reader
}
\usage{
chunk.tapply(input, FUN, ..., sep = "\t", CH.MERGE = rbind, CH.MAX.SIZE = 33554432)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input}{
}
  \item{FUN}{
}
  \item{\dots}{
}
  \item{sep}{
}
  \item{CH.MERGE}{
}
  \item{CH.MAX.SIZE}{
}
}
\details{
}
\value{

}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}